Introducing the Board
The Board is a 2D matrix of cells. It contains various exits and
maintains at all times the shortest path to the closest exit from any
cell. Some of the cells can be marked damaged (obstacle cells) which
impact the paths from cells to exits.

The functionality of the Board is evolving.